CI: attempt to silence a MSVC warning
Reviewd-by: Antonin Décimo <antonin@tarides.com>

+++ testsuite/tests/lib-bigarray-2/bigarrcstub.c
@@ -42,7 +42,8 @@ void printtab(double tab[DIMX][DIMY])
 value c_filltab(value unit)
 {
   filltab();
-  return caml_ba_alloc_dims(CAML_BA_FLOAT64 | CAML_BA_C_LAYOUT,
+  // the (int)CAML_BA_... cast is intended to avoid a MSVC warning (C5287)
+  return caml_ba_alloc_dims((int)CAML_BA_FLOAT64 | CAML_BA_C_LAYOUT,
                             2, ctab, (intnat)DIMX, (intnat)DIMY);
 }
